wasm v0.0.1 WASM.Binary

Defines an atom-and-tuple IR for the Binary Format. Gives functions for encoding the nodes, and typespecs to (partially) validate the tree structure.

Note: This module is bare and can output invalid code. See these modules for more high-level usage and further validation of WASM in Elixir:

WASM

WASM.Module

WASM.Validation

Main Concepts

encode(node)

wasm_node

wasm_value

wasm_type

wasm_instr

Types

wasm_block()

wasm_block() :: {:block, wasm_resulttype, [wasm_instr]}

wasm_const()

wasm_const() :: {:const, wasm_valtype, wasm_value}

wasm_elemtype()

wasm_elemtype() :: :elemtype

wasm_f32()

wasm_f32() :: {:f32, float_32}

32-bit floating-point.

wasm_f64()

wasm_f64() :: {:f64, float_64}

64-bit floating point.

wasm_functype()

wasm_functype() :: {:functype, wasm_vec(wasm_valtype), wasm_vec(wasm_valtype)}

wasm_globaltype()

wasm_globaltype() :: {:globaltype, :const | :var, wasm_valtype}

wasm_i32()

wasm_i32() :: {:i32, signed_32_integer}

Uninterpreted 32-bit integer (not signed or unsigned).

wasm_i64()

wasm_i64() :: {:i64, signed_64_integer}

Uninterpreted 64-bit integer (not signed or unsigned).

wasm_if()

wasm_if ::
  {:if, wasm_resulttype, [wasm_instr]} |
  {:if, wasm_resulttype, [wasm_instr], [wasm_instr]}

wasm_instr()

wasm_instr ::
  wasm_unreachable |
  wasm_nop |
  wasm_block |
  wasm_loop |
  wasm_if

wasm_limits()

wasm_limits() :: {:limits, non_neg_range | non_neg_integer}

wasm_load()

wasm_load() :: {:load, wasm_valtype, wasm_memarg}

wasm_load16_s()

wasm_load16_s() :: {:load16_s, :i32 | :i64, wasm_memarg}

wasm_load16_u()

wasm_load16_u() :: {:load16_u, :i32 | :i64, wasm_memarg}

wasm_load32_s()

wasm_load32_s() :: {:load32_u, :i64, wasm_memarg}

wasm_load32_u()

wasm_load32_u() :: {:load32_s, :i64, wasm_memarg}

wasm_load8_s()

wasm_load8_s() :: {:load8_s, :i32 | :i64, wasm_memarg}

wasm_load8_u()

wasm_load8_u() :: {:load8_s, :i32 | :i64, wasm_memarg}

wasm_loop()

wasm_loop() :: {:loop, wasm_resulttype, [wasm_instr]}

wasm_memarg()

wasm_memarg() :: {:memarg, unsigned_32_integer, unsigned_32_integer}

wasm_memtype()

wasm_memtype() :: {:memtype, wasm_limits}

wasm_name()

wasm_name() :: binary

Unicode name (UTF-8 encoded)

{:name, "foobar"}

wasm_node()

wasm_node() :: wasm_value | wasm_type | wasm_instr

Any WASM node.

wasm_nop()

wasm_nop() :: :nop

wasm_resulttype()

wasm_resulttype() :: {:resulttype, [wasm_valtype]}

wasm_s32()

wasm_s32() :: {:s32, signed_32_integer}

Signed 32-bit integer.

wasm_s64()

wasm_s64() :: {:s64, signed_64_integer}

Signed 64-bit integer.

wasm_store()

wasm_store() :: {:store, wasm_valtype, wasm_memarg}

wasm_store16()

wasm_store16() :: {:store16, :i32 | :i64, wasm_memarg}

wasm_store32()

wasm_store32() :: {:store32, :i64, wasm_memarg}

wasm_store8()

wasm_store8() :: {:store8, :i32 | :i64, wasm_memarg}

wasm_tabletype()

wasm_tabletype() :: {:tabletype, wasm_elemtype, wasm_limits}

wasm_type()

wasm_type ::
  wasm_valtype |
  wasm_resulttype |
  wasm_functype |
  wasm_limits |
  wasm_memtype |
  wasm_tabletype |
  wasm_globaltype

wasm_u32()

wasm_u32() :: {:u32, unsigned_32_integer}

Unsigned 32-bit integer.

wasm_u64()

wasm_u64() :: {:u64, unsigned_64_integer}

Unsigned 64-bit integer.

wasm_unreachable()

wasm_unreachable() :: :unreachable

wasm_valtype()

wasm_valtype() :: :i32 | :i64 | :f32 | :f64

wasm_value()

wasm_value ::
  wasm_u32 |
  wasm_u64 |
  wasm_s32 |
  wasm_s64 |
  wasm_i32 |
  wasm_i64 |
  wasm_f32 |
  wasm_f64 |
  wasm_vec |
  wasm_name

Any value node

This includes:

wasm_s32, wasm_u32, wasm_i32: 32-bit signed, unsigned, and uninterpreted integers.

Integers: Signed, unsigned, and uninterpreted, with 32-bit and 64-bit

Floating-Point: 32-bit and 64-bit

Vectors: A fixed sequence of values (or any nodes)

Names: A UTF-8 name

wasm_vec()

wasm_vec() :: wasm_vec(wasm_node)

Any vector (type of the vector itself).

wasm_vec(type)

wasm_vec(type) :: {:vec, [type]}

Vector of a specified type.

# Give the type 
wasm_vec(wasm_u32)

# Symbolizes something like:
{:vec, [
  {:u32, 1},
  {:u32, 2},
  # ...
]}

Functions

encode(node)

encode(wasm_node) :: binary

Encodes a node into its Binary Format.

TODO: Examples tests