wasm v0.0.1 WASM.Binary

Creates binary from a tree that closely resembles the WASM spec. This does not validate the code, because it only implements the rules from the “Binary Format” section in the spec. See WASM.Validation for that layer

The nodes take a simple tuple form:

```elixir
{:name, ...attributes}
```

For example, values:

```elixir
{:u32, 1000}
{:u64, 10000000000}
[;s32, -12345]
# ...
```

Or sections

```elixir
{:section, :codesec, [
  # ...
]}
```

Functions

compile(arg1)