wasm v0.0.1 WASM.Binary

Gives functions for encoding the nodes, and typespecs to (partially) validate the tree structure.

This module is only for the binary format and can output invalid code with improper use. For something more high-level, see

WASM

WASM.Module

WASM.Validation

Functions

encode(node)

encode(WASM.IR.any_node) :: binary

Encodes a node into its Binary Format.

TODO: Examples tests